- Split out layout/view classes into a 'Default' version. - Fixed animation bug when setting the tint colour for 'FieldCheckboxView'
