+++ Blockly/Tests/Code/Layout/BlockGroupLayoutTest.swift
@@ -20,13 +20,15 @@ class BlockGroupLayoutTest: XCTestCase {
 
   var _workspaceLayout: WorkspaceLayout!
   var _blockFactory: BlockFactory!
+  var _layoutFactory: LayoutFactory!
 
   // MARK: - Setup
 
   override func setUp() {
     let workspace = Workspace()
+    _layoutFactory = LayoutFactory()
     _workspaceLayout = try! WorkspaceLayout(workspace: workspace,
-      engine: LayoutEngine(), layoutBuilder: LayoutBuilder())
+      engine: LayoutEngine(), layoutBuilder: LayoutBuilder(layoutFactory: _layoutFactory))
     _blockFactory = try! BlockFactory(
       jsonPath: "all_test_blocks.json", bundle: NSBundle(forClass: self.dynamicType))
   }
@@ -34,7 +36,7 @@ class BlockGroupLayoutTest: XCTestCase {
   // MARK: - Tests
 
   func testZIndex() {
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
     blockGroupLayout.zIndex = 0
 
     // Add a bunch of block layouts
@@ -56,7 +58,7 @@ class BlockGroupLayoutTest: XCTestCase {
   }
 
   func testAppendBlockLayoutsEmpty() {
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
 
     XCTAssertEqual(0, blockGroupLayout.blockLayouts.count)
 
@@ -67,7 +69,7 @@ class BlockGroupLayoutTest: XCTestCase {
   }
 
   func testAppendBlockLayoutsNonEmpty() {
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
     blockGroupLayout.zIndex = 2
 
     // Add a bunch of block layouts
@@ -170,7 +172,7 @@ class BlockGroupLayoutTest: XCTestCase {
 
   func testReset() {
     // Create block group with a bunch of block layouts
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
     var blockLayouts = [BlockLayout]()
     for _ in 0 ..< 10 {
       blockLayouts.append(createBlockLayout())
@@ -186,7 +188,7 @@ class BlockGroupLayoutTest: XCTestCase {
   }
 
   func testMoveToWorkspacePositionForTopLevelBlockGroup() {
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
     blockGroupLayout.relativePosition = WorkspacePointMake(30, 30)
 
     // Add block group to workspace
@@ -233,7 +235,7 @@ class BlockGroupLayoutTest: XCTestCase {
   // MARK: - Helper methods
 
   func createBlockGroupLayout(numberOfBlockLayouts numberOfBlockLayouts: Int) -> BlockGroupLayout {
-    let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+    let blockGroupLayout = _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
 
     // Add a bunch of block layouts
     var blockLayouts = [BlockLayout]()
@@ -249,7 +251,7 @@ class BlockGroupLayoutTest: XCTestCase {
   func createBlockLayout() -> BlockLayout {
     let block = try! Block.Builder(name: "test").build()
     try! _workspaceLayout.workspace.addBlockTree(block)
-    let layout = BlockLayout(block: block, engine: _workspaceLayout.engine)
+    let layout = _layoutFactory.layoutForBlock(block, engine: _workspaceLayout.engine)
     block.delegate = layout
     return layout
   }
