+++ Blockly/Tests/Code/Layout/WorkspaceLayoutTest.swift
@@ -20,13 +20,15 @@ class WorkspaceLayoutTest: XCTestCase {
 
   var _workspaceLayout: WorkspaceLayout!
   var _blockFactory: BlockFactory!
+  var _layoutFactory: LayoutFactory!
 
   // MARK: - Setup
 
   override func setUp() {
     let workspace = Workspace()
+    _layoutFactory = LayoutFactory()
     _workspaceLayout = try! WorkspaceLayout(workspace: workspace,
-      engine: LayoutEngine(), layoutBuilder: LayoutBuilder())
+      engine: LayoutEngine(), layoutBuilder: LayoutBuilder(layoutFactory: _layoutFactory))
     _blockFactory = try! BlockFactory(
       jsonPath: "all_test_blocks.json", bundle: NSBundle(forClass: self.dynamicType))
   }
@@ -105,7 +107,8 @@ class WorkspaceLayoutTest: XCTestCase {
 
     // Add block group layouts
     for _ in 0 ..< 10 {
-      let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+      let blockGroupLayout =
+        _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
       allBlockGroupLayouts[blockGroupLayout.uuid] = blockGroupLayout
       _workspaceLayout.appendBlockGroupLayout(blockGroupLayout)
       // The block group's parent should be set to the _workspaceLayout now
@@ -125,7 +128,8 @@ class WorkspaceLayoutTest: XCTestCase {
 
     // Add block group layouts
     for _ in 0 ..< 10 {
-      let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+      let blockGroupLayout =
+        _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
       allBlockGroupLayouts.append(blockGroupLayout)
       _workspaceLayout.appendBlockGroupLayout(blockGroupLayout)
     }
@@ -147,7 +151,8 @@ class WorkspaceLayoutTest: XCTestCase {
 
     // Add block group layouts
     for _ in 0 ..< 10 {
-      let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+      let blockGroupLayout =
+        _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
       allBlockGroupLayouts.append(blockGroupLayout)
       _workspaceLayout.appendBlockGroupLayout(blockGroupLayout)
     }
@@ -167,7 +172,8 @@ class WorkspaceLayoutTest: XCTestCase {
   func testBringBlockGroupLayoutToFront() {
     // Add block group layouts
     for _ in 0 ..< 10 {
-      let blockGroupLayout = BlockGroupLayout(engine: _workspaceLayout.engine)
+      let blockGroupLayout =
+        _layoutFactory.layoutForBlockGroupLayout(engine: _workspaceLayout.engine)
       blockGroupLayout.zIndex = 0
       _workspaceLayout.appendBlockGroupLayout(blockGroupLayout)
     }
